@@ -1,28 +1,35 @@
 import { Directive } from '@angular/core';
-import { NG_ASYNC_VALIDATORS, Validator, AbstractControl } from '@angular/forms';
-import { Observable } from 'rxjs/Observable';
-import { debounceTime, distinctUntilChanged, first} from 'rxjs/operators';
+import {
+  NG_ASYNC_VALIDATORS,
+  Validator,
+  AbstractControl
+} from '@angular/forms';
+
+import { Observable } from 'rxjs';
+import { debounceTime, distinctUntilChanged, first } from 'rxjs/operators';
 
 @Directive({
-  // tslint:disable-next-line:directive-selector
-  selector: '[asyncEmailValidator][formControlName], [asyncEmailValidator][ngModel]',
+  selector:
+    '[appAsyncEmailValidator][formControlName], [appAsyncEmailValidator][ngModel]',
   providers: [
     {
       provide: NG_ASYNC_VALIDATORS,
-      useExisting: AsyncEmailValidatorDirective, 
+      useExisting: AsyncEmailValidatorDirective,
       multi: true
     }
   ]
 })
 export class AsyncEmailValidatorDirective implements Validator {
-  validate(c: AbstractControl): Promise<{ [key: string]: any}> | Observable < {[key: string]: any}> {
-    // return this.validateEmailPromise(c.value);
-    return this.validateEmailObservable(c.value)
-      .pipe(
-        debounceTime(1000),
-        distinctUntilChanged(),
-        first()
-      );
+  validate(
+    c: AbstractControl
+  ): Promise<{ [key: string]: any }> | Observable<{ [key: string]: any }> {
+    return this.validateEmailPromise(c.value);
+    // return this.validateEmailObservable(c.value)
+    //   .pipe(
+    //     debounceTime(1000),
+    //     distinctUntilChanged(),
+    //     first()
+    //   );
   }
 
   private validateEmailPromise(email: string) {
@@ -31,7 +38,7 @@ export class AsyncEmailValidatorDirective implements Validator {
         if (email === 'existsemail@example.com') {
           resolve({
             asyncEmailInvalid: true
-          })
+          });
         } else {
           resolve(null);
         }
@@ -42,12 +49,10 @@ export class AsyncEmailValidatorDirective implements Validator {
   private validateEmailObservable(email: string) {
     return new Observable(observer => {
       if (email === 'existsemail@example.com') {
-        observer.next({asyncEmailInvalid: true});
+        observer.next({ asyncEmailInvalid: true });
       } else {
         observer.next(null);
       }
     });
   }
-
-
 }
