@@ -2,12 +2,12 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
+import { AppRoutingModule } from './app-routing.module';
 import { ValidatorsModule } from './validators/validators.module';
 
 import { AppComponent } from './app.component';
-import { SignupFormComponent } from './template-driven-forms/signup-form';
+import { SignupFormComponent } from './template-driven-forms/signup-form/signup-form.component';
 import { SignupReactiveFormComponent } from './reactive-forms/signup-reactive-form/signup-reactive-form.component';
-
 
 @NgModule({
   declarations: [
@@ -19,9 +19,10 @@ import { SignupReactiveFormComponent } from './reactive-forms/signup-reactive-fo
     BrowserModule,
     FormsModule,
     ReactiveFormsModule,
-    ValidatorsModule
+    ValidatorsModule,
+    AppRoutingModule
   ],
   providers: [],
   bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {}
