@@ -1,5 +1,10 @@
 import { Component, OnInit } from '@angular/core';
-import { FormGroup, FormControl, FormBuilder, Validators } from '@angular/forms';
+import {
+  FormGroup,
+  FormControl,
+  FormBuilder,
+  Validators
+} from '@angular/forms';
 
 import { User } from './../../models/user';
 import { CustomValidators } from './../../validators';
@@ -10,14 +15,23 @@ import { CustomValidators } from './../../validators';
   styleUrls: ['./signup-reactive-form.component.css']
 })
 export class SignupReactiveFormComponent implements OnInit {
-  countries: Array<string> = ['Ukraine', 'Armenia', 'Belarus', 'Hungary', 'Kazakhstan', 'Poland', 'Russia'];
+  countries: Array<string> = [
+    'Ukraine',
+    'Armenia',
+    'Belarus',
+    'Hungary',
+    'Kazakhstan',
+    'Poland',
+    'Russia'
+  ];
   user: User = new User();
-
   userForm: FormGroup;
+  placeholder = {
+    email: 'Email (required)',
+    phone: 'Phone'
+  };
 
-  constructor(
-    private fb: FormBuilder
-  ) { }
+  constructor(private fb: FormBuilder) {}
 
   ngOnInit() {
     this.buildForm();
@@ -27,21 +41,30 @@ export class SignupReactiveFormComponent implements OnInit {
   save() {
     // Form model
     console.log(this.userForm);
-    // Form value
+    // Form value w/o disabled controls
     console.log(`Saved: ${JSON.stringify(this.userForm.value)}`);
+    // Form value w/ disabled controls
+    console.log(`Saved: ${JSON.stringify(this.userForm.getRawValue())}`);
   }
 
-  setNotification(notifyVia: string) {
+  onSetNotification(notifyVia: string) {
     const phoneControl = this.userForm.get('phone');
     const emailControl = this.userForm.get('email');
 
     if (notifyVia === 'text') {
       phoneControl.setValidators(Validators.required);
       emailControl.clearValidators();
-    }
-    else {
-      emailControl.setValidators( [Validators.required, Validators.pattern('[a-z0-9._%+-]+@[a-z0-9.-]+')]);
+      this.placeholder.email = 'Email';
+      this.placeholder.phone = 'Phone (required)';
+    } else {
+      emailControl.setValidators([
+        Validators.required,
+        Validators.pattern('[a-z0-9._%+-]+@[a-z0-9.-]+'),
+        Validators.email
+      ]);
       phoneControl.clearValidators();
+      this.placeholder.email = 'Email (required)';
+      this.placeholder.phone = 'Phone';
     }
     phoneControl.updateValueAndValidity();
     emailControl.updateValueAndValidity();
@@ -78,7 +101,11 @@ export class SignupReactiveFormComponent implements OnInit {
       ],
       email: [
         '',
-        [Validators.required, Validators.pattern('[a-z0-9._%+-]+@[a-z0-9.-]+')]
+        [
+          Validators.required,
+          Validators.pattern('[a-z0-9._%+-]+@[a-z0-9.-]+'),
+          Validators.email
+        ]
       ],
       phone: '',
       notification: 'email',
@@ -96,17 +123,10 @@ export class SignupReactiveFormComponent implements OnInit {
     });
   }
 
-private patchFormValues() {
+  private patchFormValues() {
     this.userForm.patchValue({
       firstName: 'Vitaliy',
       lastName: 'Zhyrytskyy'
     });
   }
-
-
 }
-
-
-
-
-
