@@ -4,16 +4,8 @@ import { CommonModule } from '@angular/common';
 import { AsyncEmailValidatorDirective, ServiceLevelDirective } from '.';
 
 @NgModule({
-  imports: [
-    CommonModule
-  ],
-  declarations: [
-    AsyncEmailValidatorDirective,
-    ServiceLevelDirective
-  ],
-  exports: [
-    AsyncEmailValidatorDirective,
-    ServiceLevelDirective
-  ]
+  imports: [CommonModule],
+  declarations: [AsyncEmailValidatorDirective, ServiceLevelDirective],
+  exports: [AsyncEmailValidatorDirective, ServiceLevelDirective]
 })
-export class ValidatorsModule { }
+export class ValidatorsModule {}
